+++ app/views/scheduling/polls/_availability_grid.html.erb
