+++ app/views/scheduling/polls/_optimal_slots.html.erb
