+++ app/views/scheduling/polls/_statistics.html.erb
