+++ app/views/scheduling/polls/_time_slot_fields.html.erb
