+++ app/views/scheduling/polls/index.html.erb
