+++ app/views/scheduling/polls/new.html.erb
