+++ app/views/scheduling/public/show.html.erb
