+++ app/views/scheduling/public/thank_you.html.erb
